+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -49,6 +49,12 @@
 * Add new key roles: `CommitteeColdKey` and `CommitteeHotKey`
 * Remove `ConstitutionalDelegCert`. Instead it now lives in `cardano-ledger-shelley` as
   `GenesisDelegCert`
+* Add `StakeCredentials` and two helper functions: `toStakeCredentials` and
+  `domRestrictedStakeCredentials`
+
+### `testlib`
+
+* Add `genValidUMapWithCreds` and `uniformSubset`
 
 ## 1.2.0.0